+++ src/api/objects/toot.ts
@@ -90,9 +90,7 @@ type TootCache = {
 
 export const JUST_MUTING = "justMuting";  // Used in the filter settings to indicate that the user is just muting this toot
 export const UNKNOWN = "unknown";
-const MAX_CONTENT_PREVIEW_CHARS = 110;
 const MAX_ID_IDX = 2;
-const MIN_CHARS_FOR_LANG_DETECT = 8;
 const BSKY_BRIDGY = 'bsky.brid.gy';
 const HASHTAG_LINK_REGEX = /<a href="https:\/\/[\w.]+\/tags\/[\w]+" class="[-\w_ ]*hashtag[-\w_ ]*" rel="[a-z ]+"( target="_blank")?>#<span>[\w]+<\/span><\/a>/i;
 const HASHTAG_PARAGRAPH_REGEX = new RegExp(`^<p>(${HASHTAG_LINK_REGEX.source} ?)+</p>`, "i");
@@ -450,7 +448,7 @@ export default class Toot implements TootObj {
      * @returns {string}
      */
     contentShortened(maxChars?: number): string {
-        maxChars ||= MAX_CONTENT_PREVIEW_CHARS;
+        maxChars ||= config.toots.maxContentPreviewChars;
         let content = this.contentString();
         content = replaceHttpsLinks(content);
 
@@ -458,8 +456,8 @@ export default class Toot implements TootObj {
         if (content.length == 0) {
             let mediaType = this.attachmentType ? `${this.attachmentType}` : "empty";
             content = `<${capitalCase(mediaType)} post by ${this.author.describe()}>`;
-        } else if (content.length > MAX_CONTENT_PREVIEW_CHARS) {
-            content = `${content.slice(0, MAX_CONTENT_PREVIEW_CHARS)}...`;
+        } else if (content.length > maxChars) {
+            content = `${content.slice(0, maxChars)}...`;
         }
 
         return content;
@@ -712,8 +710,8 @@ export default class Toot implements TootObj {
     private determineLanguage(): void {
         const text = this.contentStripped();
 
-        // if (this.isUsersOwnToot() || text.length < MIN_CHARS_FOR_LANG_DETECT) {
-        if (text.length < MIN_CHARS_FOR_LANG_DETECT) {
+        // if (this.isUsersOwnToot() || text.length < config.toots.minCharsForLanguageDetect) {
+        if (text.length < config.toots.minCharsForLanguageDetect) {
             this.language ??= config.locale.defaultLanguage;
             return;
         }
@@ -725,7 +723,7 @@ export default class Toot implements TootObj {
 
         // If there's nothing detected log a warning (if text is long enough) and set language to default
         if ((tinyLD.languageAccuracies.length + langDetector.languageAccuracies.length) == 0) {
-            if (text.length > (MIN_CHARS_FOR_LANG_DETECT * 2)) {
+            if (text.length > (config.toots.minCharsForLanguageDetect * 2)) {
                 repairLogger.warn(`no language detected`, langLogObj);
             }
 
@@ -764,7 +762,7 @@ export default class Toot implements TootObj {
         }
 
         if (this.language) {
-            if (text.length > (2 * MIN_CHARS_FOR_LANG_DETECT)) {
+            if (text.length > (2 * config.toots.minCharsForLanguageDetect)) {
                 logTrace(`No guess good enough to override language "${this.language}" for "${text}"`);
             }
         } else {