+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # NEXT RELEASE
+* Move `minCharsForLanguageDetect` and `maxContentPreviewChars` to `TootsConfig`
 
 ### v0.69.11
 (Demo App) Go back to logging out after NetworkErrors
 
-### v0.69.10 - v0.69.9
+### v0.69.9 - v0.69.10
 * Stop requiring all `FetchParams` to have a value
 * Make `TheAlgorithm.weightPresets` into a static variable